updated cirq example
