+++ README.md
@@ -43,7 +43,7 @@ The package is still under development, to use it for the time being you can clo
 ```bash
 cd path/to/project/
 pip install -r requirements_core.txt
-pip install .
+pip install .[cirq]
 ``` 
 You only need numpy to use the core functionality, the other packages that gets installed are for visualization of the graphs (matplotlib and networkx) and the circuits (google's Cirq). The latter will be removed as a requirement soon.
