+++ dynamic_qcnn/__init__.py
@@ -1,12 +1,7 @@
-from .cirq_qcnn import (
-    Qcnn as Qcnn_cirq,
-    convert_graph_to_circuit_cirq,
-)
 from .core import Qconv, Qdense, Qpool, Qfree, Qcnn, Qmotifs, Qmotif
 from .utils import plot_qcnn_graphs, plot_graph, pretty_cirq_plot
 
 __all__ = [
-    "Qcnn_cirq",
     "Qcnn",
     "Qconv",
     "Qdense",
@@ -14,7 +9,6 @@ __all__ = [
     "Qfree",
     "Qmotif",
     "Qmotifs",
-    "convert_graph_to_circuit_cirq",
     "plot_qcnn_graphs",
     "plot_graph",
     "pretty_cirq_plot",
