+++ dynamic_qcnn/cirq/__init__.py
@@ -1,0 +1,3 @@
+from .cirq_qcnn import (
+    convert_graph_to_circuit_cirq,
+)
