+++ setup.cfg
@@ -11,6 +11,10 @@ install_requires =
     numpy
     networkx
     matplotlib
-    tensorflow-quantum
     ipykernel
     jupyterlab
+
+[options.extras_require]
+cirq = cirq
+pennylane = pennylane
+
